Fix #1399 LanguageClientAware DiagnosticProvider
